\name{cropNames-methods}
\docType{methods}
\alias{cropNames-methods}
\alias{cropNames}
\title{ Extract lists of crop names  }
\description{
Methods to extract lists of crop names
}
\section{Methods}{
\describe{

\item{object = "FarmParameters"}{ Get a string vector of crop names defined in the Parameters object }

\item{object = "FarmRepresentation"}{ Get a string vector of crop names defined in the \code{FarmRepresentation} object }
}}
\keyword{methods}
