\name{document-methods}
\docType{methods}
\alias{document-methods}
\alias{document}
\title{ Methods to extract java references to document objects}
\description{
Extract the underlying java xml document representing a \code{Parameters} object
}
\section{Methods}{
\describe{

\item{params = "Parameters"}{ Extract java references to xml document objects }
}}
\keyword{methods}
