\name{eo-methods}
\docType{methods}
\alias{eo-methods}
\alias{eo}
\title{ Methods for Function eo in Package `farmR' }
\description{
Extract values of the "enterprise output" from \code{FarmRepresentation} objects
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation"}{ Get the Enterprise output of a FarmRepresentation Object }
}}
\keyword{methods}
