\name{guts-methods}
\docType{methods}
\alias{guts-methods}
\alias{guts}
\title{ Methods for Function guts in Package `farmR' }
\description{
Spill the guts of the Farm or Composite farm all over the screen
}
\section{Methods}{
\describe{

\item{object = "FarmRepresentation"}{ Dump very detailed information about the internals of the model to the screen }
}}
\keyword{methods}
