\name{isSolved-methods}
\docType{methods}
\alias{isSolved-methods}
\alias{isSolved}
\title{ Methods for Function isSolved in Package `farmLP' }
\description{
Determine whether
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation"}{ Returns 1 if the farm has been solved 0 if not}
}}
\keyword{methods}
