\name{landscapeColSums}
\alias{landscapeColSums}
\title{Summarize a list of SpatialPointsDataFrame objects}
\description{
    Returns a matrix by taking colsums from a list of spatial data frames with the same data columns
    Useful for summarizing trends in cropping or bird data across a set of simulated values
}
\usage{
  landscapeColSums(datalist)
}
\arguments{
  \item{datalist}{A list of SpatialPointsDataFrame objects with the same data columns}
}
\value{
  A Matrix of column sums
}
#\seealso{}
#\examples{}
\keyword{}
