\name{objectiveNames-methods}
\docType{methods}
\alias{objectiveNames-methods}
\alias{objectiveNames}
\title{  Methods for Function objectiveNames in Package `farmR' }
\description{
Get a list of the names of all the objectives defined in a FarmRepresentation or ObjectiveParameters object
}
\section{Methods}{
\describe{

\item{object = "FarmRepresentation"}{ Returns a list of the objectives }

\item{object = "ObjectiveParameters"}{ Returns a list of the objectives }
}}
\keyword{methods}
