\name{profit-methods}
\docType{methods}
\alias{profit-methods}
\alias{profit}
\title{ Methods for Function profit in Package `farmLP' }
\description{
Get the profit value for a solved farm
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation"}{ Return the maximum solved profit }
}}
\keyword{methods}
