\name{SBFactories}
\Rdversion{1.1}
\alias{SBFactories}
\docType{data}
\title{
Locations of Sugarbeet factories in the UK
}
\description{
SpatialPointsDataFrame with the locations of Sugarbeet factories in the UK
}
\usage{data(SBFactories)}
\format{
A SpatialPointsDataFrame with the locations and names of sugarbeet factories
}



\examples{
data(SBFactories)
}
\keyword{datasets}
