\name{setInputCost-methods}
\docType{methods}
\alias{setInputCost-methods}
\alias{setInputCost}
\title{ Methods for Function setInputCost in Package `farmLP' }
\description{
Set the cost of named inputs
}
\section{Methods}{
\describe{

\item{farm = "FarmRepresentation", inputName = "character", inputCost = "numeric"}{ Set the cost of a named input. }
}}
\keyword{methods}
